Remove restorable interface
